+++ core/components/managerbuttons/src/Processors/Group/Remove.php
@@ -9,16 +9,30 @@ class Remove extends ProcessorBase
 {
     public function process()
     {
-        $id = (int) $this->getProperty('id');
-        /** @var ButtonGroup|null $group */
-        $group = $this->modx->getObject(ButtonGroup::class, $id);
-        if (!$group) {
-            return $this->failure($this->modx->lexicon('managerbuttons_err_group_nf'));
+        $ids = $this->intList('ids');
+        $single = (int) $this->getProperty('id');
+        if ($single > 0) {
+            $ids[] = $single;
         }
-        if (!$this->service->removeGroup($group)) {
+        $ids = array_values(array_unique(array_filter($ids)));
+        if ($ids === []) {
+            return $this->failure($this->modx->lexicon('managerbuttons_err_ns'));
+        }
+
+        $removed = 0;
+        foreach ($ids as $id) {
+            /** @var ButtonGroup|null $group */
+            $group = $this->modx->getObject(ButtonGroup::class, $id);
+            if ($group && $this->service->removeGroup($group)) {
+                $removed++;
+            }
+        }
+        if ($removed === 0) {
             return $this->failure($this->modx->lexicon('managerbuttons_err_remove'));
         }
 
-        return $this->success($this->modx->lexicon('managerbuttons_group_removed'));
+        return $this->success($this->modx->lexicon('managerbuttons_group_removed'), [
+            'removed' => $removed,
+        ]);
     }
 }