+++ core/components/managerbuttons/src/Processors/Widget/GetList.php
@@ -6,6 +6,11 @@ use ManagerButtons\Processors\ProcessorBase;
 
 class GetList extends ProcessorBase
 {
+    public function checkPermissions()
+    {
+        return $this->modx->user && $this->modx->user->hasSessionContext('mgr');
+    }
+
     public function process()
     {
         $groupId = (int) $this->getProperty('group_id', 0);